+++ Geometry.h
@@ -5,10 +5,10 @@
 
 class Geometry {
 public:
-	static bool get_intersection_circle(double*, Coordinates, std::pair<Coordinates*, Coordinates*>);
+	static bool get_intersection_circle(int*, Coordinates, std::pair<Coordinates*, Coordinates*>);
 	static void up_right_left(int&, int&, int, int, int, int, bool);
-	static bool get_intersection(double*, double*, double*); 
-	static void get_equation(double*, Coordinates, Coordinates);
+	static bool get_intersection(int*, int*, double*); 
+	static void get_equation(int*, Coordinates, Coordinates);
 };
 
 #endif
